[refactor]Reconstruct episodic memory
